carmatch-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
pydantic[email]>=2.0
pydantic-settings>=2.0
email-validator>=2.0
python-jose[cryptography]>=3.3
bcrypt>=4.0
python-multipart>=0.0.9
